static_cast<> for Nongeos/
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@3071 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Nongeos/spot_light.cpp
@@ -136,7 +136,10 @@ void spot_light :: obj_setupCallback(t_class *classPtr)
   class_addmethod(classPtr, reinterpret_cast<t_method>(&spot_light::lightParamMessCallback),
                   gensym("paramlist"), A_FLOAT, A_FLOAT, A_FLOAT, A_NULL);
 }
-void spot_light :: lightParamMessCallback(void *data, t_floatarg linAtt, t_floatarg cutoff, t_floatarg exponent)
+void spot_light :: lightParamMessCallback(void *data, 
+					  t_floatarg linAtt, 
+					  t_floatarg cutoff, 
+					  t_floatarg exponent)
 {
-  GetMyClass(data)->lightParamMess((float)linAtt, (float)cutoff, (float)exponent);
+  GetMyClass(data)->lightParamMess(linAtt, cutoff, exponent);
 }
